refactor: 상품 등록 DTO 수정 - product -> auction 도메인 이동
